requirements-dev.txt
# Development and Testing Dependencies for Excel AI Assistant

# Testing Framework
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
pytest-xdist>=3.5.0  # For parallel test execution

# Code Quality
black>=24.0.0
flake8>=7.0.0
pylint>=3.0.0
mypy>=1.8.0
isort>=5.13.0

# Type stubs
types-requests>=2.31.0
pandas-stubs>=2.1.0

# Pre-commit hooks
pre-commit>=3.6.0

# Testing utilities
faker>=22.0.0  # For generating test data
freezegun>=1.4.0  # For mocking datetime
responses>=0.24.0  # For mocking HTTP requests

# Coverage visualization
coverage[toml]>=7.4.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
